spin-world.v1/backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
blessed==1.20.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
croniter==3.0.3
diff-match-patch==20241021
Django==5.1.3
django-adminlte-ui==2.1.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==4.3.3
django-picklefield==3.2
django-timezone-field==7.0
django-unfold==0.40.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
geoip2==4.8.1
idna==3.8
kombu==5.4.2
maxminddb==2.6.2
multidict==6.0.5
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
redis==3.5.3
requests==2.32.3
schema==0.7.7
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
tablib==3.7.0
twilio==9.2.4
types-python-dateutil==2.9.0.20241003
tzdata==2024.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
